sync-payments/requirements.txt
python-dotenv
stripe
schedule
beautifulsoup4
